Feat/creat views (#2)
* feat: adicionando controllers

* feat: adicionando widgets

* feat: adicionando telas

+++ lib/data/models/mileage_details_model.dart
@@ -1,9 +1,9 @@
 import 'package:passagem_express/data/models/flight_baggage_model.dart';
 
 class MileageDetails {
-  final int adulto;
-  final int crianca;
-  final int bebe;
+  final double adulto;
+  final double crianca;
+  final double bebe;
   final bool executivo;
   final double taxaEmbarque;
   final FlightBaggage limiteBagagem;
